Move applyDecision to Rules
