+++ __tests__/state.test.ts
@@ -1,6 +1,5 @@
-import { Player, Slot } from '../src/models';
-import { createInitialState, StateManager } from '../src/state';
-import { Zone } from '../src/types';
+import { Player } from '../src/models';
+import { createInitialState } from '../src/state';
 
 const sig: Player = {
     name: 'Sig',
@@ -18,22 +17,5 @@ describe('createInitialState', () => {
             return side.player;
         });
         expect(statePlayers).toEqual([sig, marla]);
-    });
-});
-
-describe('StateManager.applyDecision', () => {
-    it('should throw if the selection was not available', () => {
-        const state = createInitialState(sig, marla);
-        const stateManager = new StateManager(state);
-        const slot: Slot = {
-            zone: Zone.ENEMY_ARSENAL,
-            index: 0,
-        };
-        state.currentDeed.decisions.push({
-            label: 'top-level',
-            availableSlots: [],
-            selectedSlots: [],
-        });
-        expect(() => stateManager.applyDecision([slot])).toThrowError();
     });
 });
